# src/services/inventory/docs/swagger.yaml
definitions:
  contracts.CreateClusterRequest:
    properties:
      description:
        type: string
      host:
        type: string
      name:
        type: string
      password:
        type: string
      port:
        type: integer
      user:
        type: string
    required:
    - description
    - host
    - name
    - password
    - port
    - user
    type: object
...
  dto.CreateExchangeDto:
    properties:
      Arguments:
        additionalProperties: true
        type: object
      Durable:
        type: boolean
      Internal:
        type: boolean
      Name:
        type: string
      Type:
        type: string
    type: object
...
  dto.QueueImportRequest:
    properties:
      QueueName:
        type: string
      Type:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /{clusterId}/exchange:
    get:
      consumes:
      - application/json
      description: List all exchanges from cluster
      parameters:
      - description: Cluster id from where retrieve exchanges
        in: path
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: List all exchanges from cluster
      tags:
      - Exchange
    post:
      consumes:
      - application/json
      description: List all exchanges from cluster
      parameters:
      - description: Cluster id from where retrieve exchanges
        in: path
        name: clusterId
        required: true
        type: integer
      - description: Request
        in: body
        name: CreateExchangeDto
        required: true
        schema:
          $ref: '#/definitions/dto.CreateExchangeDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: number
      summary: List all exchanges from cluster
      tags:
      - Exchange
...
  /{clusterId}/queue/{queueId}:
    get:
      consumes:
      - application/json
      description: Retrieve a queue from cluster
      parameters:
      - description: Cluster id
        in: path
        name: clusterId
        required: true
        type: integer
      - description: Queue id
        in: path
        name: queueId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Retriave a queue from cluster
      tags:
      - Queue
  /{clusterId}/queue/import:
    post:
      consumes:
      - application/json
      description: Import existing queue from cluster or creater another one
      parameters:
      - description: Cluster id from where retrieve users
        in: path
        name: clusterId
        required: true
        type: integer
      - description: Request
        in: body
        name: QueueImportRequest
        required: true
        schema:
          $ref: '#/definitions/dto.QueueImportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Import or create queue
      tags:
      - Queue
  /{clusterId}/queue/queuesfromcluster:
    get:
      consumes:
      - application/json
      description: Retrieve all users that exist on rabbit cluster. Event if it its
        registered in ostern
      parameters:
      - description: Cluster id from where retrieve users
        in: path
        name: clusterId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Retrieve all users from rabbitmq cluster
      tags:
      - Queue
...
  /cluster:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of cluster that the user has access
      parameters:
      - default: 1
        in: query
        name: PageNumber
        required: true
        type: integer
      - default: 10
        in: query
        name: PageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Retrieve a list of rabbitmq clusters registered
      tags:
      - Cluster
    post:
      consumes:
      - application/json
      description: Create a new <b>RabbitMQ</b> cluster. The credential provider must
        be valid and the cluster operational
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/contracts.CreateClusterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
      summary: Register a new RabbitMQ Cluster
      tags:
      - Cluster
...
swagger: "2.0"